Move IndexStatistics to its own src file
