+++ src/main/scala/com/microsoft/hyperspace/index/IndexCollectionManager.scala
@@ -97,7 +97,7 @@ class IndexCollectionManager(
   override def indexes: DataFrame = {
     import spark.implicits._
     getIndexes()
-      .filter(!_.state.equals(Constants.States.DOESNOTEXIST))
+      .filter(!_.state.equals(DOESNOTEXIST))
       .map(IndexStatistics(spark, _))
       .toDF()
       .select(
@@ -167,146 +167,4 @@ object IndexCollectionManager {
       IndexLogManagerFactoryImpl,
       IndexDataManagerFactoryImpl,
       FileSystemFactoryImpl)
-}
-
-/**
- * Case class representing index metadata and index statistics from latest index version.
- *
- * @param name Index name.
- * @param indexedColumns Indexed columns.
- * @param includedColumns Included columns.
- * @param numBuckets Number of buckets.
- * @param schema Index schema json.
- * @param indexLocation Path to parent directory containing index files for all versions.
- * @param state Index state.
- * @param kind Index kind.
- * @param hasLineage Lineage enabled on index.
- * @param numIndexFiles Total number of index files.
- * @param sizeIndexFiles Total size of index files.
- * @param numSourceFiles Total number of source data files.
- * @param sizeSourceFiles Total size of source data files.
- * @param numAppendedFiles Total number of appended source data files.
- * @param sizeAppendedFiles Total size of appended source data files.
- * @param numDeletedFiles Total number of deleted source data files.
- * @param sizeDeletedFiles Total size of deleted source data files.
- * @param indexContentPaths Path(s) to directories containing index files for latest version.
- */
-private[hyperspace] case class IndexStatistics(
-    name: String,
-    indexedColumns: Seq[String],
-    includedColumns: Seq[String],
-    numBuckets: Int,
-    schema: String,
-    indexLocation: String,
-    state: String,
-    kind: String,
-    hasLineage: Boolean,
-    numIndexFiles: Int,
-    sizeIndexFiles: Long,
-    numSourceFiles: Int,
-    sizeSourceFiles: Long,
-    numAppendedFiles: Int,
-    sizeAppendedFiles: Long,
-    numDeletedFiles: Int,
-    sizeDeletedFiles: Long,
-    indexContentPaths: Seq[String])
-
-private[hyperspace] object IndexStatistics {
-  val INDEX_SUMMARY_COLUMNS: Seq[String] = Seq(
-    "name",
-    "indexedColumns",
-    "includedColumns",
-    "numBuckets",
-    "schema",
-    "indexLocation",
-    "state")
-
-  def apply(
-      spark: SparkSession,
-      entry: IndexLogEntry,
-      extended: Boolean = false): IndexStatistics = {
-    if (extended) {
-      IndexStatistics(
-        entry.name,
-        entry.derivedDataset.properties.columns.indexed,
-        entry.derivedDataset.properties.columns.included,
-        entry.numBuckets,
-        entry.derivedDataset.properties.schemaString,
-        indexDirPath(entry),
-        entry.state,
-        entry.derivedDataset.kind,
-        entry.hasLineageColumn,
-        entry.content.fileInfos.size,
-        entry.content.fileInfos.map(_.size).sum,
-        entry.sourceFileInfoSet.size,
-        entry.sourceFileInfoSet.map(_.size).sum,
-        entry.appendedFiles.size,
-        entry.appendedFiles.map(_.size).sum,
-        entry.deletedFiles.size,
-        entry.deletedFiles.map(_.size).sum,
-        getIndexContentDirectoryPaths(entry))
-    } else {
-      IndexStatistics(
-        entry.name,
-        entry.derivedDataset.properties.columns.indexed,
-        entry.derivedDataset.properties.columns.included,
-        entry.numBuckets,
-        entry.derivedDataset.properties.schemaString,
-        indexDirPath(entry),
-        entry.state,
-        "",
-        false,
-        0,
-        0L,
-        0,
-        0L,
-        0,
-        0L,
-        0,
-        0L,
-        Nil)
-    }
-  }
-
-  /**
-   * Extract top-most index directories which contain existing index files for
-   * the latest version of index. When refreshing index, depending on the mode,
-   * index files for the latest version of index may reside in multiple directories.
-   * This function extracts paths to top-level directories which
-   * contain those index files.
-   *
-   * @param entry Index log entry.
-   * @return List of directory paths containing index files for latest index version.
-   */
-  private def getIndexContentDirectoryPaths(entry: IndexLogEntry): Seq[String] = {
-    var root = entry.content.root
-    var prefix = entry.content.root.name
-    while (root.subDirs.size == 1 &&
-           !root.subDirs.head.name.startsWith(IndexConstants.INDEX_VERSION_DIRECTORY_PREFIX)) {
-      prefix += s"${root.subDirs.head.name}/"
-      root = root.subDirs.head
-    }
-
-    root.subDirs.map(d => s"$prefix${d.name}")
-  }
-
-  /**
-   * This method extracts the most top-level (or top-most) index directory which
-   * has either
-   * - at least one leaf file, or
-   * - more than one subdirectories, or
-   * - no files and no subdirectories (this case will not happen for real index scenarios).
-   *
-   * @param entry Index log entry.
-   * @return Path to the first leaf directory starting from the root.
-   */
-  private def indexDirPath(entry: IndexLogEntry): String = {
-    var root = entry.content.root
-    var indexDirPath = new Path(entry.content.root.name)
-    while (root.files.isEmpty && root.subDirs.size == 1) {
-      root = root.subDirs.head
-      indexDirPath = new Path(indexDirPath, root.name)
-    }
-    indexDirPath.toString
-  }
 }
